moved RasterHistogram.svelte
