+++ src/components/controls/RasterHistogram.svelte
